Cleaned Mark environment a little

+++ utility/envirogen/markenvironment.h
@@ -12,23 +12,23 @@ public:
 
 
 private:
-    float RandFloat();
-    quint8 add_and_limit(int oldvalue, int color, float dist, float maxdist, float tightness);
+    double RandFloat();
+    quint8 add_and_limit(int oldvalue, int color, double dist, double maxdist, double tightness);
     void ReadSettings();
-    float objxpos[MAXOBJ];
-    float objypos[MAXOBJ];
-    float objxvel[MAXOBJ];
-    float objyvel[MAXOBJ];
-    float objsize[MAXOBJ];
-    float objsizevel[MAXOBJ];
-    float objcolours[MAXOBJ][3];
-    float objcolvel[MAXOBJ][3];
-    float objtightness[MAXOBJ];
-    float objtightvel[MAXOBJ];
+    double objxpos[MAXOBJ];
+    double objypos[MAXOBJ];
+    double objxvel[MAXOBJ];
+    double objyvel[MAXOBJ];
+    double objsize[MAXOBJ];
+    double objsizevel[MAXOBJ];
+    double objcolours[MAXOBJ][3];
+    double objcolvel[MAXOBJ][3];
+    double objtightness[MAXOBJ];
+    double objtightvel[MAXOBJ];
     int objectcount;
-    float maxsize, minsize;
-    float maxvel,maxsizevel,maxcolvel,maxtightvel, maxtight, mintight, speedfactor;
-    float veltweak, sizetweak, coltweak, tighttweak;
+    double maxsize, minsize;
+    double maxvel, maxsizevel, maxcolvel, maxtightvel, maxtight, mintight, speedfactor;
+    double veltweak, sizetweak, coltweak, tighttweak;
     int iter_to_accel;
     int iter_reset;
 };
